Fix unapproved matched import transaction handling
When an imported transaction is matched with an existing transaction but not
yet approved, the YNAB API returns two copies of the transaction (the original
and the import), but the account balance only reflects one of those
transactions (since the YNAB UI shows it as a single transaction with the
"link" icon next to it).  However, this tool created a difference transaction
for the unapproved matched import (in addition to the user-created matching
transaction), and then had to create a cancelling adjustment transaction to
bring the difference account balance back into line.  One the match is
approved, YNAB deletes the imported transaction, and we would "delete" (change
to 0 amount) the difference transaction and create another adjustment
transaction that cancels out the original one.  Everything ends up being in
balance, but a bunch of confusing extra transactions were created (and the rate
adjustments category would be out of whack until the match is approved or
unmatched).

To fix, we now ignore (actually treat the same as deleted) any unapproved
matched import transaction.