Handle when user deletes a difference transaction
If a user deleted a difference transaction and then updated the foreign
transaction that created, it would result in a YNAB API error when this tool
attempted to update a difference transaction that no longer exists.

Now, we detect when the user deleted a difference transaction and remove the
internal record, which will cause this tool to create a new difference
transaction rather than try to update the deleted transaction.

This also reverts c4116db42922e5854c975833a0959de6c841403e,
since it solves the the same problem more correctly.